requirements.txt
numpy
pandas
matplotlib
Pillow
torch
torchvision
tqdm
logging
shutil
datetime
os
time
cv2
detectron2

